requirements.txt
# Core dependencies
roboflow
supervision
opencv-python-headless
streamlit
scipy
scikit-learn
polars
pandas
numpy
pyyaml
python-dotenv
plotly

# Training dependencies
torch
torchvision
transformers<5.0.0  # RF-DETR requires transformers < 5.0.0
timm
pycocotools
tensorboard
pillow
mlflow
rfdetr>=1.3.0
huggingface_hub
# byte-track - removed, package name may be different or not needed